backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
sentence-transformers
faiss-cpu
numpy
scipy
scikit-learn
google-generativeai
httpx
beautifulsoup4
readability-lxml
PyPDF2
pdfminer.six
python-dotenv
rapidfuzz
pytest

